feat: add history sidebar to folder details page
