+++ resources/js/components/drawers/FolderDetailsSidebarDrawer.vue
@@ -1,16 +1,10 @@
 <script setup lang="ts">
-import { defineAsyncComponent, watch } from 'vue';
-import { useFolderTabs } from '@/components/folders/FolderTabs';
-import { useAppStore } from '@/stores/AppStore.ts';
-import { storeToRefs } from 'pinia';
 import { BaseDrawer } from '@/components/cedar-ui/drawer';
 import { useDrawer } from '@aminnausin/cedar-ui';
 import { useRoute } from 'vue-router';
+import { watch } from 'vue';
 
-const { selectedSideBar } = storeToRefs(useAppStore());
-const { activeFolderTab } = useFolderTabs();
-
-const FolderSidebarAsync = defineAsyncComponent(async () => await import('@/components/panels/FolderSidebar.vue'));
+import FolderDetailsSidebar from '@/components/panels/FolderDetailsSidebar.vue';
 
 const drawer = useDrawer();
 const route = useRoute();
@@ -23,12 +17,6 @@ watch(
 
 <template>
     <BaseDrawer>
-        <Suspense v-if="selectedSideBar === 'folders'">
-            <FolderSidebarAsync :url-suffix="`${['details', activeFolderTab?.name].filter(Boolean).join('/')}`" />
-
-            <template #fallback>
-                <SidebarSkeleton />
-            </template>
-        </Suspense>
+        <FolderDetailsSidebar />
     </BaseDrawer>
 </template>
